devtools events for same event should be processed sequentially
Some events (like HeapProfiler.addHeapSnapshotChunk) provides data that has to be used in the order it was produced (otherwise it aggregated result does not make sense). Current implementation does not guarantee that event consumer will be called for event in the same order event has been triggered. This has been changed